the very last fix

+++ lib/tasks/whiteboard.rake
@@ -7,11 +7,12 @@ namespace :whiteboard do
       dir.each do |file|
         filename = args[:dir] + file
         next if file.match /^\./
+        next if File.exists?(RAILS_ROOT + '/public/images/' + file.downcase)
         topic_id = file.downcase.match(/^t(\d+)_/)[1] if file.downcase.match(/^t(\d+)_/)
         title = file.match(/^(.+)\./).nil? ? 'whiteboard file' : file.match(/^(.+)\./)[1]
         if file.downcase.match /(png|jpg|gif)$/
           File.rename(filename, RAILS_ROOT + '/public/images/' + file.downcase)
-          Photo.create!(:body => image_url(file.downcase), :topic_ids => topic_id, :title => title)
+          Photo.create!(:body => '/images/' + file.downcase), :topic_ids => topic_id, :title => title)
         elsif file.downcase.match /txt$/
           f = File.new(filename, 'r')
           Article.create!(:source => '/txt/' + file.downcase, :body => f.read, :topic_ids => topic_id, :title => title)
